--- intercept/doc.go
@@ -1,2 +1,0 @@
-// Package intercept banana.
-package intercept
