+++ middleware/doc.go
@@ -1,0 +1,2 @@
+// Package middleware banana.
+package middleware
